Organize user utilities by moving getUserRoleAndName function to userUtils file
Related to #1521
